Split executor and validator commands
