+++ nomos-cli/Cargo.toml
@@ -11,12 +11,10 @@ clap = { version = "4", features = ["derive"] }
 executor-http-client = { path = "../clients/executor-http-client" }
 hex = "0.4.3"
 kzgrs-backend = { path = "../nomos-da/kzgrs-backend" }
+nomos-core = { path = "../nomos-core/chain-defs" }
+nomos-node = { path = "../nodes/nomos-node" }
 reqwest = { version = "0.12", features = ["json"] }
 serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-serde_yaml = "0.9"
-thiserror = "1.0"
 tokio = { version = "1", features = ["sync"] }
-tokio-stream = "0.1"
 tracing = "0.1"
 tracing-subscriber = "0.3"
