+++ nomos-cli/src/cmds/mod.rs
@@ -1,4 +1,5 @@
 pub mod executor;
+pub mod validator;
 
 // std
 // crates
@@ -9,11 +10,11 @@ use clap::Subcommand;
 pub enum Command {
     /// Send data to the executor for encoding and dispersal.
     Disseminate(executor::Disseminate),
-    Retrieve(executor::Retrieve),
+    Retrieve(validator::Retrieve),
 }
 
 impl Command {
-    pub fn run(&self) -> Result<(), Box<dyn std::error::Error>> {
+    pub fn run(self) -> Result<(), Box<dyn std::error::Error>> {
         match self {
             Command::Disseminate(cmd) => cmd.run(),
             Command::Retrieve(cmd) => cmd.run(),
